refactor(publications): remove validation code from Collection.publish
