+++ lib/data_aggregator/records/publication/changes/publish_records.ex
@@ -16,29 +16,20 @@ defmodule DataAggregator.Records.Publication.Changes.PublishRecords do
     Changeset.before_transaction(changeset, &publish_records(&1, ctx), append?: true)
   end
 
-  defp publish_records(%Changeset{data: %{channel: :fast_track} = original_publication} = changeset, %{
-         actor: actor,
-         tenant: tenant
-       }) do
+  defp publish_records(%Changeset{data: %{} = original_publication} = changeset, %{actor: actor, tenant: tenant}) do
     publication = Ash.load!(original_publication, [:collection])
 
-    case Collection.publish(publication, actor: actor, authorize?: false, tenant: tenant) do
+    case publish_or_validate(publication, actor, tenant) do
       {:ok, publication} -> add_success(changeset, publication, tenant)
       {:error, error} -> add_error(changeset, error, publication)
     end
   end
 
-  defp publish_records(%Changeset{data: %{channel: :validation} = original_publication} = changeset, %{
-         actor: actor,
-         tenant: tenant
-       }) do
-    publication = Ash.load!(original_publication, [:collection])
+  defp publish_or_validate(%{channel: :fast_track} = publication, actor, tenant),
+    do: Collection.publish(publication, actor: actor, authorize?: false, tenant: tenant)
 
-    case Collection.send_validation(publication, actor: actor, authorize?: false, tenant: tenant) do
-      {:ok, publication} -> add_success(changeset, publication, tenant)
-      {:error, error} -> add_error(changeset, error, publication)
-    end
-  end
+  defp publish_or_validate(%{channel: :validation} = publication, actor, tenant),
+    do: Collection.validate(publication, actor: actor, authorize?: false, tenant: tenant)
 
   defp add_error(changeset, error, publication) do
     Logger.warning("Error while publishing records: #{inspect(error)}")
