+++ lib/data_aggregator/records/publication/changes/set_collection_publishing_before_transaction.ex
@@ -2,7 +2,7 @@ defmodule DataAggregator.Records.Publication.Changes.SetCollectionPublishingBefo
   @moduledoc """
   Set the collection to fast_track_publishing before we start the publication itself.
   If the collection is not idle, we will not allow the publication to start. If the channel
-  is validation, we do not update the collection state. This is done in the Collection.validate
+  is validation, we do not update the collection state. This is done in the Collection.start_validations
   action.
   """
