+++ lib/data_aggregator_web/live/collection_live/record/index.ex
@@ -1048,7 +1048,7 @@ defmodule DataAggregatorWeb.CollectionLive.Record.Index do
   end
 
   defp create_and_enqueue(collection, query, _count_query, :validation, actor) do
-    Collection.validate(collection, query, actor: actor, tenant: collection)
+    Collection.start_validations(collection, query, actor: actor, tenant: collection)
   end
 
   defp apply_action(socket, :index, _params) do
